Grouped scripts containing generic helper functions into utils.py for clarity.
